+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 * Remove deprecated support for restricting uploads to certain mime types.
+* Remove deprecated support for `forUploads`.
 
 ### Fixed